refactor: moved NetworkManager into test because it's only used in tests
